AI/app.py
"""
HANGUL Learning App - Flask AI Backend

Services:
- Camera object detection (YOLOv8)
- Pronunciation scoring (Azure Speech Services)
- Text-to-speech (Google Cloud)
"""

from flask import Flask, request, jsonify
from flask_cors import CORS
import base64
import numpy as np
from PIL import Image
import io

# Initialize Flask app
app = Flask(__name__)
CORS(app)

# ========================
# ROUTES
# ========================

@app.route('/health', methods=['GET'])
def health_check():
    """Health check endpoint"""
    return jsonify({
        'status': 'ok',
        'service': 'HANGUL Flask AI Backend',
        'version': '0.1.0'
    })

@app.route('/api/detect-camera', methods=['POST'])
def detect_camera():
    """
    Camera object detection endpoint
    
    Request:
        {
            "image": "base64_image_string",
            "canvasSize": {"width": 640, "height": 480}
        }
    
    Response:
        {
            "objects": [
                {
                    "name": "cup",
                    "korean": "컵",
                    "romanization": "keop",
                    "confidence": 0.95
                }
            ]
        }
    """
    try:
        data = request.json
        image_base64 = data.get('image')
        
        if not image_base64:
            return jsonify({'error': 'No image provided'}), 400
        
        # Decode base64 image
        image_data = base64.b64decode(image_base64)
        image = Image.open(io.BytesIO(image_data))
        
        # TODO: Run YOLOv8 detection here
        # from ultralytics import YOLO
        # model = YOLO('yolov8n.pt')
        # results = model(image)
        
        # Placeholder response
        objects = [
            {
                'name': 'cup',
                'korean': '컵',
                'romanization': 'keop',
                'confidence': 0.95
            }
        ]
        
        return jsonify({'objects': objects}), 200
        
    except Exception as e:
        return jsonify({'error': str(e)}), 500

@app.route('/api/pronunciation-score', methods=['POST'])
def score_pronunciation():
    """
    Pronunciation scoring endpoint
    
    Request:
        {
            "audio": "base64_audio_string",
            "word": "apple",
            "userId": 1
        }
    
    Response:
        {
            "accuracyScore": 85,
            "fluencyScore": 80,
            "completenessScore": 90,
            "overallScore": 85,
            "feedback": "Good pronunciation!"
        }
    """
    try:
        data = request.json
        audio_base64 = data.get('audio')
        word = data.get('word')
        
        if not audio_base64 or not word:
            return jsonify({'error': 'Missing audio or word'}), 400
        
        # TODO: Implement Azure Speech Service integration or similar
        # from azure.cognitiveservices.speech import speechconfig, SpeechRecognizer
        
        # Placeholder response
        result = {
            'accuracyScore': 85,
            'fluencyScore': 80,
            'completenessScore': 90,
            'overallScore': 85,
            'feedback': 'Good pronunciation!',
            'phoneticErrors': []
        }
        
        return jsonify(result), 200
        
    except Exception as e:
        return jsonify({'error': str(e)}), 500

if __name__ == '__main__':
    print('🚀 Flask AI Backend running on port 5001')
    app.run(host='0.0.0.0', port=5001, debug=True)
