AI/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# ML & Computer Vision
ultralytics==8.2.0
torch==2.3.0
torchvision==0.18.0
opencv-python==4.8.0.76
pillow==10.0.0
numpy==1.24.0

# Audio Processing
librosa==0.10.0
soundfile==0.12.1

# Azure Cognitive Services
azure-cognitiveservices-speech==1.31.0

# Google Cloud
google-cloud-texttospeech==2.14.0

# Utilities
requests==2.31.0
